Address PR review feedback on IFDS/followX
- Move IfdsSummaryCache into its own file; shared tabulation types/
  helpers (PathEdge, SummaryRecord, CallerRecord, TokenKey, IfdsStep +
  subtypes, ifdsStepsOf, ifdsTokensMatch) become internal instead of
  file-private so both files can use them.
- Split eogSteps into eogStepsBackward/eogStepsForward helpers instead
  of one function handling both directions.
- Simplify the public KDoc on followDFGEdgesUntilHitNodes/
  followEOGEdgesUntilHitNodes to describe params/purpose; move the
  delegation-guard rationale into a code comment, including why this
  only covers the node-set (MAY) variants and not the path-collecting
  ones (a summary edge is reused across callers and doesn't by itself
  carry a witness path back to the start).
- Drop "frontier" from the public follow*UntilHitNodes docs in favor of
  plain "node"/"PDG node"/"CDG node".
- Note the single-direction-per-solve limitation on
  IfdsReachingSourcesSolver.backward.
- Drop a stray CBOM-specific example from IfdsSummaryCache's KDoc.
- Add a short comment to expandPure.
