+++ cpg-core/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/IfdsReachingSources.kt
@@ -116,29 +116,30 @@ fun Node.ifdsReachingSourcesEOG(
 ): Set<Node> = ifdsReachingSources(direction, k, predicate)
 
 /** A single interprocedural step produced by the successor oracle. */
-private sealed interface IfdsStep
+internal sealed interface IfdsStep
 
 /** An intraprocedural move to [target] (no push/pop). */
-private class IntraStep(val target: Node) : IfdsStep
+internal class IntraStep(val target: Node) : IfdsStep
 
 /**
  * A call (push) step: from [callSite] we enter a callee whose entry node is [calleeEntry], pushing
  * [token] (the call site(s), matched by identity) onto the conceptual call stack.
  */
-private class CallStep(val callSite: Node, val calleeEntry: Node, val token: List<Call>) : IfdsStep
+internal class CallStep(val callSite: Node, val calleeEntry: Node, val token: List<Call>) :
+    IfdsStep
 
 /**
  * A return (pop) step: we leave the current procedure by popping [token]; the caller resumes at
  * each node in [targets].
  */
-private class ReturnStep(val token: List<Call>, val targets: List<Node>) : IfdsStep
+internal class ReturnStep(val token: List<Call>, val targets: List<Node>) : IfdsStep
 
 /**
  * Identity-based key over a call-string token (a list of [Call]s). Two tokens are equal iff they
  * reference the same [Call] objects in the same order (object identity), so structurally-equal but
  * distinct call sites are never conflated.
  */
-private class TokenKey(val calls: List<Call>) {
+internal class TokenKey(val calls: List<Call>) {
     override fun equals(other: Any?): Boolean {
         if (other !is TokenKey) return false
         if (calls.size != other.calls.size) return false
@@ -154,7 +155,7 @@ private class TokenKey(val calls: List<Call>) {
 }
 
 /** A tabulated summary of a callee: popping [token] out of the callee resumes at [target]. */
-private class SummaryRecord(val token: TokenKey, val target: Node) {
+internal class SummaryRecord(val token: TokenKey, val target: Node) {
     override fun equals(other: Any?): Boolean {
         return other is SummaryRecord && token == other.token && target === other.target
     }
@@ -166,14 +167,14 @@ private class SummaryRecord(val token: TokenKey, val target: Node) {
  * A caller waiting on a callee's summary: the caller lived in context [callerCtx] (its own entry
  * node, or `null` for the empty-stack ROOT region), pushed at [callSite] with [token].
  */
-private class CallerRecord(val callerCtx: Node?, val callSite: Node, val token: List<Call>)
+internal class CallerRecord(val callerCtx: Node?, val callSite: Node, val token: List<Call>)
 
 /**
  * A tabulated path edge: node [node] is reachable within the procedure context identified by [ctx]
  * (its entry node, or `null` for the empty-stack ROOT region). Equality is by object identity of
  * both components.
  */
-private class PathEdge(val ctx: Node?, val node: Node) {
+internal class PathEdge(val ctx: Node?, val node: Node) {
     override fun equals(other: Any?): Boolean {
         return other is PathEdge && ctx === other.ctx && node === other.node
     }
@@ -192,6 +193,9 @@ private class IfdsReachingSourcesSolver(
      */
     val cache: IfdsSummaryCache? = null,
 ) {
+    // A single solver instance analyzes exactly one AnalysisDirection: mixing forward and backward
+    // steps within the same query (or genuinely bidirectional analysis) is not supported and would
+    // require running two separate solves. No current caller needs this.
     private val backward = direction is Backward
     private val graph = direction.graphToFollow
 
@@ -302,7 +306,7 @@ private class IfdsReachingSourcesSolver(
  * the given traversal direction ([backward]) and sub-[graph]. It depends only on the graph
  * structure, never on a predicate, which is exactly why the resulting summaries can be cached.
  */
-private fun ifdsStepsOf(node: Node, backward: Boolean, graph: GraphToFollow): List<IfdsStep> {
+internal fun ifdsStepsOf(node: Node, backward: Boolean, graph: GraphToFollow): List<IfdsStep> {
     return when (graph) {
         GraphToFollow.DFG -> dfgSteps(node, backward)
         GraphToFollow.EOG -> eogSteps(node, backward)
@@ -329,34 +333,48 @@ private fun dfgSteps(node: Node, backward: Boolean): List<IfdsStep> {
 }
 
 private fun eogSteps(node: Node, backward: Boolean): List<IfdsStep> {
-    if (backward) {
-        if (node is Call && node.invokes.isNotEmpty()) {
-            val steps = node.invokeEdges.map { inv -> CallStep(node, inv.end, listOf(node)) }
-            if (steps.isNotEmpty()) return steps
-        } else if (node is Function) {
-            val steps =
-                node.calledByEdges.mapNotNull { inv ->
-                    val call = inv.start as? Call ?: return@mapNotNull null
-                    ReturnStep(listOf(call), call.prevEOG.toList())
-                }
-            if (steps.isNotEmpty()) return steps
-        }
-        return node.prevEOGEdges.map { IntraStep(it.start) }
-    } else {
-        if (node is Call && node.invokes.isNotEmpty()) {
-            val steps = node.invokeEdges.map { inv -> CallStep(node, inv.end, listOf(node)) }
-            if (steps.isNotEmpty()) return steps
-        } else if (node is Return || node.nextEOG.isEmpty()) {
-            val fn = (node as? Function) ?: node.firstParentOrNull<Function>()
-            val steps =
-                fn?.calledByEdges?.mapNotNull { inv ->
-                    val call = inv.start as? Call ?: return@mapNotNull null
-                    ReturnStep(listOf(call), call.nextEOG.toList())
-                } ?: emptyList()
-            if (steps.isNotEmpty()) return steps
-        }
-        return node.nextEOGEdges.map { IntraStep(it.end) }
+    return if (backward) eogStepsBackward(node) else eogStepsForward(node)
+}
+
+/**
+ * Backward EOG successor oracle: a [Call] with invokes always pushes into the callee; otherwise a
+ * [Function] node is treated as the callee's boundary and always pops back to its callers; any
+ * other node just walks its own [Node.prevEOGEdges].
+ */
+private fun eogStepsBackward(node: Node): List<IfdsStep> {
+    if (node is Call && node.invokes.isNotEmpty()) {
+        val steps = node.invokeEdges.map { inv -> CallStep(node, inv.end, listOf(node)) }
+        if (steps.isNotEmpty()) return steps
+    } else if (node is Function) {
+        val steps =
+            node.calledByEdges.mapNotNull { inv ->
+                val call = inv.start as? Call ?: return@mapNotNull null
+                ReturnStep(listOf(call), call.prevEOG.toList())
+            }
+        if (steps.isNotEmpty()) return steps
     }
+    return node.prevEOGEdges.map { IntraStep(it.start) }
+}
+
+/**
+ * Forward EOG successor oracle: a [Call] with invokes always pushes into the callee; a [Return] or
+ * a leaf node (no further [Node.nextEOG]) pops back through every call site invoking its enclosing
+ * [Function]; any other node just walks its own [Node.nextEOGEdges].
+ */
+private fun eogStepsForward(node: Node): List<IfdsStep> {
+    if (node is Call && node.invokes.isNotEmpty()) {
+        val steps = node.invokeEdges.map { inv -> CallStep(node, inv.end, listOf(node)) }
+        if (steps.isNotEmpty()) return steps
+    } else if (node is Return || node.nextEOG.isEmpty()) {
+        val fn = (node as? Function) ?: node.firstParentOrNull<Function>()
+        val steps =
+            fn?.calledByEdges?.mapNotNull { inv ->
+                val call = inv.start as? Call ?: return@mapNotNull null
+                ReturnStep(listOf(call), call.nextEOG.toList())
+            } ?: emptyList()
+        if (steps.isNotEmpty()) return steps
+    }
+    return node.nextEOGEdges.map { IntraStep(it.end) }
 }
 
 /**
@@ -364,7 +382,7 @@ private fun eogSteps(node: Node, backward: Boolean): List<IfdsStep> {
  * superset). Otherwise the (single-frame) tokens must reference the same call site(s) by identity.
  * Shared by the query solver and the pure summary cache.
  */
-private fun ifdsTokensMatch(callerToken: List<Call>, returnToken: List<Call>, k: Int): Boolean {
+internal fun ifdsTokensMatch(callerToken: List<Call>, returnToken: List<Call>, k: Int): Boolean {
     if (k <= 0) return true
     if (callerToken.size != returnToken.size) return false
     // Compare the top min(k, size) frames by identity. Tokens are single-frame in practice, so any
@@ -381,174 +399,3 @@ private fun ifdsTokensMatch(callerToken: List<Call>, returnToken: List<Call>, k:
  */
 private fun AnalysisDirection.sameShapeAs(other: AnalysisDirection): Boolean =
     this::class == other::class && this.graphToFollow == other.graphToFollow
-
-/**
- * A per-pass, cross-query cache for [ifdsReachingSources]. It exploits two facts that hold while a
- * pass runs:
- * 1. The DFG/EOG **graph is immutable** during the pass, so *predicate-free* balanced callee
- *    summaries ("enter this callee, pop this token, resume at this node") are stable graph facts
- *    that can be tabulated once and reused by every query.
- * 2. The only thing a query's [predicate] changes is where the frontier stops. A callee whose
- *    transitive callee-subgraph contains **no sink** can therefore never trigger frontier pruning,
- *    so its pure summary is exactly what any query would compute — we replay it and skip the
- *    callee.
- *
- * A callee is **dirty** if its function (transitively, up the call graph) contains a sink. Sinks
- * are registered by the owning pass via [markSink] as overlays / state entries are added to nodes
- * (that is what a [predicate] like the CBOM `getNodesBasedOnDfg` tests). Dirtiness is monotonic
- * (sinks are only ever added) and propagated to all callers, so once a callee is dirty it stays
- * dirty and the solver falls back to exact, predicate-aware tabulation for it — preserving results
- * bit-for-bit.
- *
- * ## Soundness contract
- * The cache is only sound for queries whose sink set is a subset of the nodes reported through
- * [markSink]. Callers guarantee this by installing the cache on a [Component] (see
- * [Component.ifdsSummaryCache]) **only** for the duration of a pass whose queries have exactly that
- * shape, and by [markSink]-ing every node that could satisfy such a query's predicate (including a
- * one-time scan of pre-existing overlays). Any query that finds no cache runs the exact uncached
- * tabulation.
- *
- * Instances are used from a single pass that executes its queries sequentially; no internal locking
- * is required.
- *
- * @param direction The traversal this cache's pure summaries were tabulated for. A query with a
- *   different [AnalysisDirection] shape must not reuse them (enforced in [ifdsReachingSources]).
- */
-class IfdsSummaryCache(val direction: AnalysisDirection) {
-    private val backward = direction is Backward
-    private val graph = direction.graphToFollow
-
-    // ---- predicate-free ("pure") tabulation state, shared/incremental across all queries ----
-    private val pureSummaries: MutableMap<Node, MutableSet<SummaryRecord>> = IdentityHashMap()
-    private val pureCallers: MutableMap<Node, MutableList<CallerRecord>> = IdentityHashMap()
-    private val pureVisited = HashSet<PathEdge>()
-    private val pureWorklist = ArrayDeque<PathEdge>()
-
-    /** Callee entries whose pure summary set is fully tabulated (complete). */
-    private val pureComplete: MutableSet<Node> = Collections.newSetFromMap(IdentityHashMap())
-
-    /** Callee entries seeded during the in-progress drain; all become complete once it finishes. */
-    private val seededThisDrain: MutableSet<Node> = Collections.newSetFromMap(IdentityHashMap())
-
-    // ---- dirtiness ----
-    private val dirty: MutableSet<Function> = Collections.newSetFromMap(IdentityHashMap())
-    private val functionOfCache = IdentityHashMap<Node, Function?>()
-
-    /**
-     * Replays the complete pure summary of the clean callee entered at [calleeEntry], invoking
-     * [emit] with each caller-resume target whose token balances [callerToken] under [k]. Computes
-     * the pure summary on first use.
-     */
-    internal fun replayCleanCallee(
-        calleeEntry: Node,
-        callerToken: List<Call>,
-        k: Int,
-        emit: (Node) -> Unit,
-    ) {
-        ensurePure(calleeEntry)
-        pureSummaries[calleeEntry]?.toList()?.forEach { sum ->
-            if (ifdsTokensMatch(callerToken, sum.token.calls, k)) emit(sum.target)
-        }
-    }
-
-    /** Ensures [calleeEntry] has a complete pure summary set, tabulating it on demand. */
-    private fun ensurePure(calleeEntry: Node) {
-        if (calleeEntry in pureComplete) return
-        seededThisDrain.clear()
-        seedPure(calleeEntry, calleeEntry)
-        seededThisDrain.add(calleeEntry)
-        while (pureWorklist.isNotEmpty()) {
-            val pe = pureWorklist.removeFirst()
-            expandPure(pe.ctx, pe.node)
-        }
-        // A full drain makes every callee context it seeded complete (recursion is bounded by
-        // summary reuse, exactly as in the query solver).
-        pureComplete.addAll(seededThisDrain)
-    }
-
-    private fun seedPure(ctx: Node?, node: Node) {
-        // No predicate: the pure tabulation treats no node as a sink.
-        if (pureVisited.add(PathEdge(ctx, node))) pureWorklist.addLast(PathEdge(ctx, node))
-    }
-
-    private fun expandPure(ctx: Node?, node: Node) {
-        for (step in ifdsStepsOf(node, backward, graph)) {
-            when (step) {
-                is IntraStep -> seedPure(ctx, step.target)
-                is CallStep -> {
-                    val callee = step.calleeEntry
-                    seededThisDrain.add(callee)
-                    pureCallers
-                        .getOrPut(callee) { mutableListOf() }
-                        .add(CallerRecord(ctx, step.callSite, step.token))
-                    seedPure(callee, callee)
-                    pureSummaries[callee]?.toList()?.forEach { sum ->
-                        if (ifdsTokensMatch(step.token, sum.token.calls, Int.MAX_VALUE)) {
-                            seedPure(ctx, sum.target)
-                        }
-                    }
-                }
-                is ReturnStep -> {
-                    if (ctx == null) {
-                        for (t in step.targets) seedPure(null, t)
-                        continue
-                    }
-                    val tokenKey = TokenKey(step.token)
-                    for (t in step.targets) {
-                        val sr = SummaryRecord(tokenKey, t)
-                        if (pureSummaries.getOrPut(ctx) { HashSet() }.add(sr)) {
-                            pureCallers[ctx]?.toList()?.forEach { cr ->
-                                if (ifdsTokensMatch(cr.token, step.token, Int.MAX_VALUE)) {
-                                    seedPure(cr.callerCtx, t)
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Registers that [node] is (now) a sink for the queries this cache serves, marking its function
-     * — and, transitively, every caller — dirty so their pure summaries are no longer reused.
-     */
-    fun markSink(node: Node) {
-        val f = functionOf(node) ?: return
-        if (!dirty.add(f)) return
-        val work = ArrayDeque<Function>()
-        work.addLast(f)
-        while (work.isNotEmpty()) {
-            val g = work.removeFirst()
-            for (call in g.calledBy) {
-                val caller = call.firstParentOrNull<Function>() ?: continue
-                if (dirty.add(caller)) work.addLast(caller)
-            }
-        }
-    }
-
-    /**
-     * True if the callee entered at [calleeEntry] may contain a sink (so its pure summary must not
-     * be reused). Unknown-function callees are treated as dirty (conservative).
-     */
-    fun isDirty(calleeEntry: Node): Boolean {
-        val f = functionOf(calleeEntry) ?: return true
-        return f in dirty
-    }
-
-    /**
-     * The [Function] a node belongs to (via AST for program nodes, via underlying for overlays).
-     */
-    private fun functionOf(node: Node): Function? {
-        if (functionOfCache.containsKey(node)) return functionOfCache[node]
-        val f =
-            when (node) {
-                is Function -> node
-                is OverlayNode ->
-                    node.underlyingNode?.let { it as? Function ?: it.firstParentOrNull<Function>() }
-                else -> node.firstParentOrNull<Function>()
-            }
-        functionOfCache[node] = f
-        return f
-    }
-}
